Add files via upload
During the transfer of the code from Google Colab to the GitHub repository, the code was damaged and the direct GitHub link stopped working, showing an Invalidation error. For this reason, the codes and all project steps are being re-run from the beginning, and every part of the code is being executed again one by one.